feat: replace Three.js CDN with npm import; convert to ES modules
Removes the Three.js CDN script tag and imports `three` from npm in
js/main.js (`import * as THREE from 'three'`).  Vite now bundles the
library, so the browser fetches a single hashed JS chunk instead of an
external CDN script.

To make this work end-to-end without losing test coverage:

* js/main.js + js/europe-map.js are now ES modules (`export {...}` in
  place of `module.exports = {...}`).
* All data files (data/locations.js, data/publications.js,
  data/projects.js, data/cv.js) are ES modules that also assign to
  globalThis, so the existing tests can still mock data via
  `global.X = mock` and the runtime classes can read bare globals.
* scripts/generate-cv.js, scripts/generate-locations.js, and
  scripts/new-project.js emit the new ESM format.
* Nested `package.json` files in `data/` and `js/` set
  `"type": "module"` so only those directories switch to ESM; CJS
  scripts and test runners are unaffected.
* Tests that exercise main.js / europe-map.js are renamed to .mjs and
  use ES `import`.  THREE mocking switches from `global.THREE = mock`
  to `__setThreeForTests(mock)` / `__resetThreeForTests()` (exported
  helpers), since the module-local THREE binding can no longer be
  swapped via globals.
* HTML pages now load a single `<script type="module" src="/js/main.js">`
  and the production bundle is fully self-contained.

207 tests pass (added one for the generator's ES module form).
`npm run build` produces a working dist/.
