+++ cv.html
@@ -107,8 +107,7 @@
     </svg>
   </button>
 
-  <script defer src="data/cv.js"></script>
-  <script defer src="js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 
 </html>
