+++ data/locations.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 /* Generated by scripts/generate-locations.js from data\locations.yaml */
-const LOCATIONS = {
+export const LOCATIONS = {
   "pins": [
     {
       "type": "lived",
@@ -1259,3 +1259,5 @@ const LOCATIONS = {
   "regions": []
 };
 
+
+if (typeof globalThis !== 'undefined') globalThis.LOCATIONS = LOCATIONS;
