+++ package.json
@@ -7,11 +7,11 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "test": "node --test test/main.node.test.js test/europe-map.test.js test/locations-generator.test.js test/new-project.test.js test/cv.test.js test/generate-cv.test.js",
-    "test:main": "node --test test/main.node.test.js",
+    "test": "node --test test/main.node.test.mjs test/europe-map.test.mjs test/locations-generator.test.js test/new-project.test.js test/cv.test.mjs test/generate-cv.test.js",
+    "test:main": "node --test test/main.node.test.mjs",
     "test:locations": "node --test test/locations-generator.test.js",
     "test:project": "node --test test/new-project.test.js",
-    "test:cv": "node --test test/cv.test.js",
+    "test:cv": "node --test test/cv.test.mjs",
     "test:generate-cv": "node --test test/generate-cv.test.js",
     "new-project": "node scripts/new-project.js",
     "generate-cv": "node scripts/generate-cv.js",
@@ -26,5 +26,8 @@
   "devDependencies": {
     "playwright": "^1.58.2",
     "vite": "^8.0.10"
+  },
+  "dependencies": {
+    "three": "^0.184.0"
   }
 }
