+++ projects.html
@@ -96,8 +96,7 @@
     </svg>
   </button>
 
-  <script defer src="data/projects.js"></script>
-  <script defer src="js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 
 </html>
