+++ projects/aroundtheworld.html
@@ -165,6 +165,6 @@ If there is one thing I took from this work, it is that the bottleneck for AR in
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
