+++ projects/audience-engagement.html
@@ -166,6 +166,6 @@ The work fed into subsequent research at Vicomtech on <strong>multi-device media
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
