+++ projects/avatech.html
@@ -142,6 +142,6 @@ AVATecH shaped my approach to every project that followed. The constraints of th
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
