+++ projects/clear-architecture.html
@@ -210,7 +210,7 @@
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 
 </html>
