+++ projects/rag-document-qa.html
@@ -136,6 +136,6 @@ No need to know which folder anything lives in — you just ask, and you get an 
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
