+++ projects/ufc-fighter-tracking.html
@@ -152,6 +152,6 @@ This was a project where I was involved <strong>end to end</strong>:
     </svg>
   </button>
 
-  <script defer src="../js/main.min.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
