+++ scripts/generate-locations.js
@@ -255,7 +255,9 @@ function toLocationsJs(locations, sourcePath) {
   );
   return `/* eslint-disable */
 /* Generated by scripts/generate-locations.js from ${sourcePath} */
-const LOCATIONS = ${payload};
+export const LOCATIONS = ${payload};
+
+if (typeof globalThis !== 'undefined') globalThis.LOCATIONS = LOCATIONS;
 `;
 }
